+++ src/main/java/lotto/service/LottoGameService.java
@@ -34,7 +34,7 @@ public class LottoGameService {
         return lottos.getLottoRanks(winningLotto, bonusNumber);
     }
 
-    public float showProfitRate(EnumMap<LottoRank, Integer> lottoRanks) {
+    public double showProfitRate(EnumMap<LottoRank, Integer> lottoRanks) {
         int profit = lottoRanks.entrySet().stream()
                 .filter(e -> e.getValue() > 0)
                 .map(e -> e.getKey().getPrizeMoney() * e.getValue())
